clarify physic module
