+++ physic/examples/test_dynamic_rect.py
@@ -1,7 +1,7 @@
 import pygame
 from pygame import Vector2, Rect
 
-from ..physic_rect import DynamicRect, PhysicRect
+from ..rect import DynamicRect, PhysicRect
 
 
 def main():
