--- physic/physic_rect/__init__.py
@@ -1,2 +1,0 @@
-from .physic_rect import DynamicRect, \
-        FloatRect, PhysicRect, CollisionResponce
