+++ physic/rect/__init__.py
@@ -1,0 +1,8 @@
+# from .physic_rect import DynamicRect, \
+#         FloatRect, PhysicRect, CollisionResponce
+
+
+from .float_rect import FloatRect
+from .collision_responce import CollisionResponce
+from .physic_rect import PhysicRect
+from .dynamic_rect import DynamicRect
